remove dc.js dependency
