+++ frontend/src/metabase/css/vendor.css
@@ -1,3 +1,2 @@
-@import "dc/dc.css";
 @import "ace-builds/css/ace.css";
 @import "tippy.js/dist/tippy.css";
