+++ frontend/src/metabase/visualizations/echarts/cartesian/model/axis.unit.spec.ts
@@ -1,0 +1,25 @@
+import { computeSplit } from "./axis";
+import type { SeriesExtents } from "./types";
+
+describe("computeSplit", () => {
+  const extents: SeriesExtents = {
+    "1": [6, 8],
+    "2": [9, 13],
+    "3": [6, 7],
+    "4": [1, 1],
+    "5": [10, 13],
+    "6": [15, 19],
+    "7": [5, 6],
+    "8": [5, 10],
+    "9": [9, 13],
+    "10": [2, 6],
+    "11": [12, 15],
+    "12": [1, 1],
+  };
+
+  it("should return the same number of series as given", () => {
+    expect(computeSplit(extents).flat()).toHaveLength(
+      Object.keys(extents).length,
+    );
+  });
+});
