+++ frontend/src/metabase/visualizations/lib/renderer_utils.js
@@ -1,94 +1,14 @@
-/// Utility functions used by both the LineAreaBar renderer and the RowRenderer
-
 import { getIn } from "icepick";
 import _ from "underscore";
 
 import { NULL_NUMERIC_VALUE } from "metabase/lib/constants";
 import { formatNullable } from "metabase/lib/formatting/nullable";
 import { parseTimestamp } from "metabase/lib/time";
-import { isNative } from "metabase-lib/v1/queries/utils/card";
 import { datasetContainsNoResults } from "metabase-lib/v1/queries/utils/dataset";
-import { isNumeric } from "metabase-lib/v1/types/utils/isa";
 
 import { dimensionIsNumeric } from "./numeric";
-import { getLineAreaBarComparisonSettings } from "./settings";
-import {
-  dimensionIsTimeseries,
-  dimensionIsExplicitTimeseries,
-} from "./timeseries";
-import { getAvailableCanvasWidth, getAvailableCanvasHeight } from "./utils";
+import { dimensionIsExplicitTimeseries } from "./timeseries";
 import { invalidDateWarning, nullDimensionWarning } from "./warnings";
-
-export function initChart(chart, element) {
-  // set the bounds
-  chart.width(getAvailableCanvasWidth(element));
-  chart.height(getAvailableCanvasHeight(element));
-  // disable animations
-  chart.transitionDuration(0);
-  // disable brush
-  if (chart.brushOn) {
-    chart.brushOn(false);
-  }
-}
-
-export function makeIndexMap(values) {
-  const indexMap = new Map();
-  for (const [index, key] of values.entries()) {
-    indexMap.set(key, index);
-  }
-  return indexMap;
-}
-
-// HACK: This ensures each group is sorted by the same order as xValues,
-// otherwise we can end up with line charts with x-axis labels in the correct order
-// but the points in the wrong order. There may be a more efficient way to do this.
-export function forceSortedGroup(group, indexMap) {
-  const sorted = group
-    .top(Infinity)
-    .sort((a, b) => indexMap.get(a.key) - indexMap.get(b.key));
-  for (let i = 0; i < sorted.length; i++) {
-    sorted[i].index = i;
-  }
-  group.all = () => sorted;
-}
-
-export function forceSortedGroupsOfGroups(groupsOfGroups, indexMap) {
-  for (const groups of groupsOfGroups) {
-    for (const group of groups) {
-      forceSortedGroup(group, indexMap);
-    }
-  }
-}
-
-/*
- * The following functions are actually just used by LineAreaBarRenderer but moved here in interest of making that namespace more concise
- */
-
-export function reduceGroup(group, key, warnUnaggregated) {
-  return group.reduce(
-    (acc, d) => {
-      if (acc == null && d[key] == null) {
-        return null;
-      } else {
-        if (acc != null && d[key] != null) {
-          warnUnaggregated();
-        }
-        return (acc || 0) + (d[key] || 0);
-      }
-    },
-    (acc, d) => {
-      if (acc == null && d[key] == null) {
-        return null;
-      } else {
-        if (acc != null && d[key] != null) {
-          warnUnaggregated();
-        }
-        return (acc || 0) - (d[key] || 0);
-      }
-    },
-    () => null,
-  );
-}
 
 export function parseXValue(xValue, options, warn) {
   const { parsedValue, warning } = memoizedParseXValue(xValue, options);
@@ -130,44 +50,6 @@ function getParseOptions({ settings, data }) {
 function canDisplayNull(settings) {
   // histograms are converted to ordinal scales, so we need this ugly logic as a workaround
   return !isOrdinal(settings) || isHistogram(settings);
-}
-
-export function getDatas({ settings, series }, warn) {
-  return series.map(({ data }) => {
-    const parseOptions = getParseOptions({ settings, data });
-
-    let rows = data.rows;
-
-    // non-ordinal dimensions can't display null values,
-    // so we filter them out and display a warning
-    if (canDisplayNull(settings)) {
-      rows = data.rows.filter(([x]) => x !== null);
-    } else if (parseOptions.isNumeric) {
-      rows = data.rows.map(row => {
-        const [x, ...rest] = row;
-        const newRow = [replaceNullValuesForOrdinal(x), ...rest];
-        newRow._origin = row._origin;
-        return newRow;
-      });
-    }
-
-    if (rows.length < data.rows.length) {
-      warn(nullDimensionWarning());
-    }
-
-    return rows.map(row => {
-      const [x, ...rest] = row;
-      const { unit } = parseOptions;
-      const xValue = parseXValue(x, parseOptions, warn);
-      const formattedXValue =
-        xValue && unit && typeof xValue.startOf === "function"
-          ? xValue.startOf(unit)
-          : xValue;
-      const newRow = [formattedXValue, ...rest];
-      newRow._origin = row._origin;
-      return newRow;
-    });
-  });
 }
 
 export function getXValues({ settings, series }) {
@@ -238,67 +120,6 @@ function parseTimestampAndWarn(value, unit) {
   return { parsedValue: m };
 }
 
-const uniqueCards = series => _.uniq(series.map(({ card }) => card.id)).length;
-
-const getMetricColumnsCount = series => {
-  const metricColumnPredicate = !isNative(series[0]?.card)
-    ? column => column.source === "aggregation"
-    : column => isNumeric(column);
-
-  return _.uniq(
-    series
-      .flatMap(({ data: { cols } }) => cols.filter(metricColumnPredicate))
-      .map(({ name }) => name),
-  ).length;
-};
-
-export function shouldSplitYAxis(
-  { settings, chartType, isScalarSeries, series },
-  datas,
-  yExtents,
-) {
-  const isSuitableChartType = !isScalarSeries && chartType !== "scatter";
-  if (!isSuitableChartType) {
-    return false;
-  }
-
-  if (!settings["graph.y_axis.auto_split"]) {
-    return false;
-  }
-
-  const isSingleCardWithSingleMetricColumn =
-    uniqueCards(series) <= 1 && getMetricColumnsCount(series) <= 1;
-
-  if (isSingleCardWithSingleMetricColumn || isStacked(settings, datas)) {
-    return false;
-  }
-
-  const hasDifferentYAxisColTypes =
-    _.uniq(series.map(s => s.data.cols[1].semantic_type)).length > 1;
-  if (hasDifferentYAxisColTypes) {
-    return true;
-  }
-
-  const columnSettings = series.map(s =>
-    getLineAreaBarComparisonSettings(settings.column(s.data.cols[1])),
-  );
-  const hasDifferentColumnSettings = columnSettings.some(s1 =>
-    columnSettings.some(s2 => !_.isEqual(s1, s2)),
-  );
-  if (hasDifferentColumnSettings) {
-    return true;
-  }
-
-  const minRange = Math.min(...yExtents.map(extent => extent[1] - extent[0]));
-  const maxExtent = Math.max(...yExtents.map(extent => extent[1]));
-  const minExtent = Math.min(...yExtents.map(extent => extent[0]));
-  const chartRange = maxExtent - minExtent;
-
-  // Note (EmmadUsmani): When the series with the smallest range is less than 5%
-  // of the chart's total range, we split the y-axis so it doesn't look too small.
-  return minRange / chartRange <= 0.05;
-}
-
 /************************************************************ PROPERTIES ************************************************************/
 
 export const isTimeseries = settings =>
@@ -310,15 +131,6 @@ export const isHistogram = settings =>
   settings["graph.x_axis.scale"] === "histogram";
 export const isOrdinal = settings =>
   settings["graph.x_axis.scale"] === "ordinal";
-export const isLine = settings => settings.display === "line";
-export const isArea = settings => settings.display === "area";
-
-// bar histograms have special tick formatting:
-// * aligned with beginning of bar to show bin boundaries
-// * label only shows beginning value of bin
-// * includes an extra tick at the end for the end of the last bin
-export const isHistogramBar = ({ settings, chartType }) =>
-  isHistogram(settings) && chartType === "bar";
 
 export const isStacked = (settings, datas) => settings["stackable.stack_type"];
 export const isNormalized = (settings, datas) =>
@@ -327,8 +139,6 @@ export const isNormalized = (settings, datas) =>
 // find the first nonempty single series
 export const getFirstNonEmptySeries = series =>
   _.find(series, s => !datasetContainsNoResults(s.data));
-export const isDimensionTimeseries = series =>
-  dimensionIsTimeseries(getFirstNonEmptySeries(series).data);
 
 function hasRemappingAndValuesAreStrings({ cols }, i = 0) {
   const column = cols[i];
@@ -348,12 +158,6 @@ export const isRemappedToString = series =>
 export const hasClickBehavior = series =>
   getIn(series, [0, "card", "visualization_settings", "click_behavior"]) !=
   null;
-
-// is this a dashboard multiseries?
-// TODO: better way to detect this?
-export const isMultiCardSeries = series =>
-  series.length > 1 &&
-  getIn(series, [0, "card", "id"]) !== getIn(series, [1, "card", "id"]);
 
 // Hack: for numeric dimensions we have to replace null values
 // with anything else since crossfilter groups merge 0 and null
