+++ frontend/src/metabase/visualizations/visualizations/CartesianChart/chart-definition-legacy.jsx
@@ -1,0 +1,157 @@
+import _ from "underscore";
+
+import { NULL_DISPLAY_VALUE } from "metabase/lib/constants";
+import { formatValue } from "metabase/lib/formatting";
+import { isEmpty } from "metabase/lib/validate";
+import { getComputedSettingsForSeries } from "metabase/visualizations/lib/settings/visualization";
+import { getFriendlyName } from "metabase/visualizations/lib/utils";
+
+// TODO: this series transformation is used only for the visualization settings computation which is excessive.
+// Replace this with defining settings models per visualization type which will contain all necessary info
+// about the dataset sufficient to compute settings defaults like series settings.
+export function transformSeries(series) {
+  const newSeries = [].concat(
+    ...series.map((s, seriesIndex) =>
+      transformSingleSeries(s, series, seriesIndex),
+    ),
+  );
+  if (_.isEqual(series, newSeries) || newSeries.length === 0) {
+    return series;
+  } else {
+    return newSeries;
+  }
+}
+
+function transformSingleSeries(s, series, seriesIndex) {
+  const { card, data } = s;
+
+  // HACK: prevents cards from being transformed too many times
+  if (data._transformed) {
+    return [s];
+  }
+
+  const { cols, rows } = data;
+  const settings = getComputedSettingsForSeries([s]);
+
+  const dimensions = (settings["graph.dimensions"] || []).filter(
+    d => d != null,
+  );
+  const metrics = (settings["graph.metrics"] || []).filter(d => d != null);
+  const dimensionColumnIndexes = dimensions.map(dimensionName =>
+    _.findIndex(cols, col => col.name === dimensionName),
+  );
+  const metricColumnIndexes = metrics.map(metricName =>
+    _.findIndex(cols, col => col.name === metricName),
+  );
+  const bubbleColumnIndex =
+    settings["scatter.bubble"] &&
+    _.findIndex(cols, col => col.name === settings["scatter.bubble"]);
+  const extraColumnIndexes =
+    bubbleColumnIndex != null && bubbleColumnIndex >= 0
+      ? [bubbleColumnIndex]
+      : [];
+
+  if (dimensions.length > 1) {
+    const [dimensionColumnIndex, seriesColumnIndex] = dimensionColumnIndexes;
+    const rowColumnIndexes = [dimensionColumnIndex].concat(
+      metricColumnIndexes,
+      extraColumnIndexes,
+    );
+
+    const breakoutValues = [];
+    const breakoutRowsByValue = new Map();
+
+    for (let rowIndex = 0; rowIndex < rows.length; rowIndex++) {
+      const row = rows[rowIndex];
+      const seriesValue = row[seriesColumnIndex];
+
+      let seriesRows = breakoutRowsByValue.get(seriesValue);
+      if (!seriesRows) {
+        breakoutRowsByValue.set(seriesValue, (seriesRows = []));
+        breakoutValues.push(seriesValue);
+      }
+
+      const newRow = rowColumnIndexes.map(columnIndex => row[columnIndex]);
+      newRow._origin = { seriesIndex, rowIndex, row, cols };
+      seriesRows.push(newRow);
+    }
+
+    return breakoutValues.map(breakoutValue => ({
+      card: {
+        ...card,
+        // if multiseries include the card title as well as the breakout value
+        name: [
+          // show series title if it's multiseries
+          series.length > 1 && card.name,
+          // always show grouping value
+          formatValue(
+            isEmpty(breakoutValue) ? NULL_DISPLAY_VALUE : breakoutValue,
+            { column: cols[seriesColumnIndex] },
+          ),
+        ]
+          .filter(n => n)
+          .join(": "),
+        originalCardName: card.name,
+        _breakoutValue: breakoutValue,
+        _breakoutColumn: cols[seriesColumnIndex],
+      },
+      data: {
+        rows: breakoutRowsByValue.get(breakoutValue),
+        cols: rowColumnIndexes.map(i => cols[i]),
+        _rawCols: cols,
+        _transformed: true,
+      },
+      // for when the legend header for the breakout is clicked
+      clicked: {
+        dimensions: [
+          {
+            value: breakoutValue,
+            column: cols[seriesColumnIndex],
+          },
+        ],
+      },
+    }));
+  } else {
+    // dimensions.length <= 1
+    const dimensionColumnIndex = dimensionColumnIndexes[0];
+    return metricColumnIndexes.map(metricColumnIndex => {
+      const col = cols[metricColumnIndex];
+      const rowColumnIndexes = [dimensionColumnIndex].concat(
+        metricColumnIndex,
+        extraColumnIndexes,
+      );
+      const name = [
+        // show series title if it's multiseries
+        series.length > 1 && card.name,
+        // show column name if there are multiple metrics or sigle series
+        (metricColumnIndexes.length > 1 || series.length === 1) &&
+          col &&
+          getFriendlyName(col),
+      ]
+        .filter(n => n)
+        .join(": ");
+
+      return {
+        card: {
+          ...card,
+          name: name,
+          originalCardName: card.name,
+          _seriesIndex: seriesIndex,
+          // use underlying column name as the seriesKey since it should be unique
+          // EXCEPT for dashboard multiseries, so check seriesIndex == 0
+          _seriesKey: seriesIndex === 0 && col ? col.name : name,
+        },
+        data: {
+          rows: rows.map((row, rowIndex) => {
+            const newRow = rowColumnIndexes.map(i => row[i]);
+            newRow._origin = { seriesIndex, rowIndex, row, cols };
+            return newRow;
+          }),
+          cols: rowColumnIndexes.map(i => cols[i]),
+          _transformed: true,
+          _rawCols: cols,
+        },
+      };
+    });
+  }
+}
