+++ frontend/src/metabase/visualizations/visualizations/ComboChart/ComboChart.tsx
@@ -1,6 +1,5 @@
 import { t } from "ttag";
 
-import LineAreaBarChart from "metabase/visualizations/components/LineAreaBarChart";
 import {
   getDefaultSize,
   getMinSize,
@@ -26,7 +25,6 @@ import type {
 Object.assign(
   ComboChart,
   getCartesianChartDefinition({
-    transformSeries: LineAreaBarChart.transformSeries,
     uiName: t`Combo`,
     identifier: "combo",
     iconName: "lineandbar",
