+++ package.json
@@ -53,7 +53,6 @@
     "d3-scale": "^3.3.0",
     "d3-shape": "^3.1.0",
     "dayjs": "^1.10.4",
-    "dc": "2.1.9",
     "diff": "^3.2.0",
     "echarts": "^5.5.0",
     "formik": "^2.4.5",
